fix: :bug: Corrección del drag&drop y aplicando modal
Se aplicó modal para la publicación del usuario. Falta estilado y algo de funcionalidad
