+++ client/src/components/PostCreate/PostCreate.module.css
@@ -60,12 +60,6 @@
   background-color: #323232;
 }
 
-.input_files_container{
-  display: flex;
-  justify-content: space-around;
-  margin-top: 10px;
-}
-
 .button_post{
   background-color: #323232;
   border-radius: 5px;
