+++ client/src/components/index.js
@@ -3,5 +3,6 @@ import { Login } from './Forms/Login/Login.jsx'
 import { SignUp } from './Forms/SignUp/SignUp'
 import { Narbar } from './Navbar/Navbar'
 import { PostCreate } from './PostCreate/PostCreate.jsx'
+import { ModalPost } from './ModalPost/ModalPost.jsx'
 
-export { Feets, Login, SignUp, Narbar, PostCreate }
+export { Feets, Login, SignUp, Narbar, PostCreate, ModalPost }
